requirements.txt
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-modeltranslation==0.19.9
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.8
multidict==6.1.0
mysqlclient==2.2.6
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stripe==10.11.0
twilio==9.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
yarl==1.11.1
social-auth-app-django==5.4.2
pyOpenSSL==24.2.1
Werkzeug==3.0.4
django_extensions==3.2.3
pillow==10.4.0
django-environ==0.11.2
drf-yasg==1.21.8
gunicorn==21.2.0
